moved utilities to separate file
