+++ src/solidity/index.ts
@@ -1,3 +1,4 @@
 // TODO(ritave): Split into seperate package
 export * from "./importresolver";
 export * from "./ethpmresolver";
+export * from "./utils";
